Refactor templating, policy routing, exit code
Add helper to concatenate several templates for a single policy. Unify
common parts of unbridged/bridged (90%).

Add rules/rules6 clauses as currently present. This is not the right
moment to get rid of policy routing.

fc-network now exits with 0 if nothing is to be changed, 1 on exception,
2 on usage error, and 3 if something has been changed / should be
changed (in no-op mode).

Re #17369